+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `client` parameter to `api.update_workload_description()` for providing an existing Beaker client,
   which avoids creating one each time the function is called.
 
+### Changed
+
+- You can now specify the `--env` option as just `--env 'NAME'` instead of `--env 'NAME=VALUE'` to take the `VALUE` from a local environment variable of that name.
+- You can now specify the `--env-secret` option as just `--env-secret 'NAME'` instead of `--env-secret 'NAME=SECRET_NAME'` to create a new secret from a local environment variable of that name.
+
 ### Fixed
 
 - Made `api.update_workload_description()` more efficient.